refactor rope
