+++ python/sglang/srt/models/longcat_flash_nextn.py
@@ -148,7 +148,6 @@ class LongcatFlashDenseDecoderLayer(nn.Module):
         )
         if _is_npu:
             self.self_attn.use_explicit_npu_interleaved_rope = True
-            self.self_attn.rotary_emb.use_explicit_npu_interleaved_rope = True
 
         self.mlp = LongcatFlashMLP(
             hidden_size=config.hidden_size,
